requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
bcrypt
pyjwt
pyjwt[crypto]